Keep daemon schemas out of the app boot path
