+++ packages/host-daemon-contract/src/protocol.ts
@@ -1,0 +1,1 @@
+export const HOST_DAEMON_PROTOCOL_VERSION = 127 as const;
