+++ packages/host-daemon-contract/src/workspace.ts
@@ -1,0 +1,24 @@
+import { z } from "zod";
+
+export const workspaceResolutionFailureCodeSchema = z.enum([
+  "path_not_found",
+  "not_git_repo",
+  "not_worktree",
+  "workspace_type_mismatch",
+  "permission_denied",
+  "unknown_environment",
+  "unknown",
+]);
+export const workspaceResolutionFailureSchema = z
+  .object({
+    code: workspaceResolutionFailureCodeSchema,
+    workspacePath: z.string().min(1),
+    message: z.string().min(1),
+  })
+  .strict();
+export type WorkspaceResolutionFailureCode = z.infer<
+  typeof workspaceResolutionFailureCodeSchema
+>;
+export type WorkspaceResolutionFailure = z.infer<
+  typeof workspaceResolutionFailureSchema
+>;
